implement round 1 feedback. clippy and fmt
